web-time-control-javascript
This is a Chinese-type clock with a simple pure black interface. The central control panel can be accessed from online web pages and can be operated with related headless browser controls such as NodeRED、Hass.

index.html defaults to local time display, countdown.html is the operation interface and code details for setting localized time countdown, the reserved font is to add more new functions in the future and add various other requirements for adaptation. This project is for personal entertainment